requirements.txt
cloudinary==1.3.1
defusedxml==0.4.1
Django==1.8.8
django-braces==1.8.1
django-endless-pagination==2.0
django-model-utils==2.4
django-star-ratings==0.5.1
django-tagging==0.4.1
Markdown==2.6.5
mysqlclient==1.3.7
oauthlib==1.0.3
PyJWT==1.4.0
python-social-auth==0.2.14
python3-openid==3.0.9
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
wheel==0.29.0
